geom/tile: rename to `geom/layout`
